+++ chroma-renderer/core/scene/ModelImporter.cpp
@@ -591,8 +591,8 @@ bool convert(const aiScene* aiscene, Scene& s)
         // tan(FOV_H/2) / screen_width = tan(FOV_V/2) / screen_height
         s.camera.aspectRatio = cam->mAspect;
         s.camera.width = 640;
-        s.camera.d = ((float)s.camera.width / 2.0f) / tan(cam->mHorizontalFOV / 2.0f);
-        s.camera.height = static_cast<int>(s.camera.width / cam->mAspect);
+        s.camera.d = ((float)s.camera.width / 2.0f) / tanf(cam->mHorizontalFOV / 2.0f);
+        s.camera.height = static_cast<int>((float)s.camera.width / cam->mAspect);
         s.camera.horizontalFOV(cam->mHorizontalFOV);
 
         // s.camera.computeUVW();
